requirements.txt
streamlit>=1.30.0
pydantic>=2.0.0
pandas>=2.0.0
openai>=1.0.0
pyyaml>=6.0.0
